third_party/blink/web_tests/wpt_internal/cobalt/HTMLVideoElementExtension/getAndSetMaxVideoCapabilities.html

<!DOCTYPE html>
<html>
<body>
  <script src=/resources/testharness.js></script>
  <script src=/resources/testharnessreport.js></script>
  <script type='module'>
    test(function(){
    var video = document.createElement('video');
    assert_equals(video.getMaxVideoCapabilities(), '', 'getMaxVideoCapabilities can retrieve unset value');
    assert_not_equals(video.getMaxVideoCapabilities(), '3', 'Assert that maxVideoCapabilities is properly updated');
    }, 'getmaxVideoCapabilities for HTMLVideoElement');

    test(function(){
    var video = document.createElement('video');
    video.src = 'testStringToSetSRC';
    assert_implements(video.setMaxVideoCapabilities, 'setMaxVideoCapabilities is validly implemented.')
    assert_throws_dom('InvalidStateError', video.setMaxVideoCapabilities('width=1920; height=1080; framerate=15;'),
                      'setMaxVideoCapabilities throws InvalidStateError');
    }, 'setMaxVideoCapabilities throws INVALID_STATE_ERROR');

    test(function(){
    var video = document.createElement('video');
    video.setMaxVideoCapabilities('width=1920; height=1080; framerate=15;');
    assert_equals(video.getMaxVideoCapabilities(), 'width=1920; height=1080; framerate=15;',
                  'setMaxVideoCapabilities properly updates maxVideoCapabilities');
    }, 'setMaxVideoCapabilities successfully updates HTMLVideoElement');
  </script>
</body>
</html>
